PDF_OCR/requirements.txt
pdf2image
pytesseract
Pillow